Optimizing file reads
