+++ medschool/Cargo.toml
@@ -23,6 +23,7 @@ clap = { version = "4", features = ["derive"] }
 tracing.workspace = true
 tracing-subscriber.workspace = true
 regex.workspace = true
+threadpool = "1.8.1"
 
 [dev-dependencies]
 rand = "0.8"
